# Improve performance of writing raw UTF-8 encoded byte arrays