The output escape table covers just 7-bits, meaning that a raw UTF-8 byte cannot be used to index into the table without a branch test for negative bytes (i.e. bytes larger than 0x7F). This extra check occurs in a tight loop and can be avoided if the lookup table were to cover all 8-bit indices.

This commit introduces ad-hoc logic in `UTF8JsonGenerator#writeUTF8String` to create an extended copy of `_outputEscapes` if necessary, writing the copy back into the field to avoid having to compute it again within the same generator instance (unless it is changed). This ad-hoc strategy was chosen as it is the least disruptive to existing code, as a larger-scale change around `CharacterEscapes` would impact public api or otherwise subtle chances for breakages.

---

Some quick and dirty JMH tests on M1 Max (arm64 ⚠) with Azul Zulu JDK 21, this shows the following numbers:

```
Benchmark                (length)  (needEscape)  (optimized)   Mode  Cnt   Score   Error   Units
JmhTest.writeUtf8String        32         first         true  thrpt   40  32,156 ± 0,084  ops/us
JmhTest.writeUtf8String        32         first        false  thrpt   40  27,936 ± 0,106  ops/us
JmhTest.writeUtf8String        32          last         true  thrpt   40  33,049 ± 0,091  ops/us
JmhTest.writeUtf8String        32          last        false  thrpt   40  29,605 ± 0,102  ops/us
JmhTest.writeUtf8String        32          none         true  thrpt   40  32,922 ± 0,192  ops/us
JmhTest.writeUtf8String        32          none        false  thrpt   40  29,654 ± 0,074  ops/us
JmhTest.writeUtf8String       256         first         true  thrpt   40   6,350 ± 0,023  ops/us
JmhTest.writeUtf8String       256         first        false  thrpt   40   4,734 ± 0,012  ops/us
JmhTest.writeUtf8String       256          last         true  thrpt   40   6,399 ± 0,018  ops/us
JmhTest.writeUtf8String       256          last        false  thrpt   40   4,759 ± 0,017  ops/us
JmhTest.writeUtf8String       256          none         true  thrpt   40   6,402 ± 0,021  ops/us
JmhTest.writeUtf8String       256          none        false  thrpt   40   4,751 ± 0,025  ops/us
JmhTest.writeUtf8String       512         first         true  thrpt   40   3,215 ± 0,030  ops/us
JmhTest.writeUtf8String       512         first        false  thrpt   40   2,478 ± 0,008  ops/us
JmhTest.writeUtf8String       512          last         true  thrpt   40   3,259 ± 0,012  ops/us
JmhTest.writeUtf8String       512          last        false  thrpt   40   2,480 ± 0,026  ops/us
JmhTest.writeUtf8String       512          none         true  thrpt   40   3,262 ± 0,013  ops/us
JmhTest.writeUtf8String       512          none        false  thrpt   40   2,486 ± 0,007  ops/us
```

This is writing buffers of length `(length)` that contain `'a'` in all positions, where `(needEscape)` being `'first'` has the first byte overwritten with `"`, or `'last'` its last byte, versus `'none'` where the buffer remains a sequence where no escapes need to be inserted.

Overall the numbers show improvements in the range of 11%–33%. I wonder if this extends to other CPU architectures, opening this PR to gauge interest in such a change. Note that this only affects `UTF8JsonGenerator#writeUTF8String` which isn't typically used, as it's more common to process from `char[]` or `String` buffers. In my use-case I already have an UTF-8 encoded `byte[]` which prompted me looking into this.

---

This logic can probably be vectorized quite nicely, that is also being done in [dotnet's JSON writer infrastructure](https://github.com/dotnet/runtime/blob/dfc2b85a00f3baf5ac52d7615bf857e6217011c8/src/libraries/System.Text.Encodings.Web/src/System/Text/Encodings/Web/OptimizedInboxTextEncoder.cs#L343).