@JoostK 2024-10-21
Sure I can add some; while looking at the existing ones I wonder what the desired testing strategy is:

1. extract both write loops to be able to compare prior state (7-bit LUT) versus new state (8-bit LUT), or
2. call into `JsonGenerator#writeUTF8String` and then running the test with and without the change applied, possibly adding `char[]` writing as a comparative benchmark.

What is the most valuable thing to have here? 1. is meaningful to compare this particular change across machines/JVMs, but 2. is more valuable to measure and compare write perf of `JsonGenerator` going forward.