@pjfanning 2024-10-21
Both sound useful - could you add both benchmarks?