@JoostK 2024-10-21
> Both sound useful - could you add both benchmarks?

I'll come up with something, probably over the coming days.