@JoostK 2024-10-27
Accidentally rebased onto `master`, unaware that this PR was targeting `2.19`. Reverted back to 2.19.

Here's the results on my MBP w/ M1 Max:

```
after:

Length    8,  none escape: (7-bit, 8-bit, JsonGenerator):  57,4 /  47,8 / 196,7 msecs
Length    8, start escape: (7-bit, 8-bit, JsonGenerator):  84,3 /  76,2 / 198,5 msecs
Length    8,   end escape: (7-bit, 8-bit, JsonGenerator):  73,8 /  70,9 / 231,6 msecs
Length   16,  none escape: (7-bit, 8-bit, JsonGenerator):  64,0 /  56,3 / 120,0 msecs
Length   16, start escape: (7-bit, 8-bit, JsonGenerator):  73,5 /  62,4 / 129,5 msecs
Length   16,   end escape: (7-bit, 8-bit, JsonGenerator):  67,7 /  59,4 / 172,5 msecs
Length   32,  none escape: (7-bit, 8-bit, JsonGenerator):  63,0 /  52,9 /  85,8 msecs
Length   32, start escape: (7-bit, 8-bit, JsonGenerator):  67,0 /  56,4 / 103,6 msecs
Length   32,   end escape: (7-bit, 8-bit, JsonGenerator):  63,3 /  54,3 / 146,5 msecs
Length  256,  none escape: (7-bit, 8-bit, JsonGenerator):  60,1 /  52,1 /  60,7 msecs
Length  256, start escape: (7-bit, 8-bit, JsonGenerator):  60,8 /  54,7 /  83,5 msecs
Length  256,   end escape: (7-bit, 8-bit, JsonGenerator):  61,9 /  51,7 / 138,7 msecs
Length  512,  none escape: (7-bit, 8-bit, JsonGenerator):  59,5 /  50,9 /  56,5 msecs
Length  512, start escape: (7-bit, 8-bit, JsonGenerator):  61,6 /  53,3 /  79,3 msecs
Length  512,   end escape: (7-bit, 8-bit, JsonGenerator):  60,8 /  50,1 / 132,9 msecs
Length 1024,  none escape: (7-bit, 8-bit, JsonGenerator):  60,2 /  50,7 /  95,5 msecs
Length 1024, start escape: (7-bit, 8-bit, JsonGenerator):  60,3 /  52,5 /  78,7 msecs
Length 1024,   end escape: (7-bit, 8-bit, JsonGenerator):  60,6 /  50,1 /  97,0 msecs
Length 8192,  none escape: (7-bit, 8-bit, JsonGenerator):  58,0 /  49,0 /  32,4 msecs
Length 8192, start escape: (7-bit, 8-bit, JsonGenerator):  59,1 /  49,0 /  38,1 msecs
Length 8192,   end escape: (7-bit, 8-bit, JsonGenerator):  58,9 /  48,9 /  34,6 msecs



before:

Length    8,  none escape: (7-bit, 8-bit, JsonGenerator):  58,8 /  45,4 / 196,7 msecs
Length    8, start escape: (7-bit, 8-bit, JsonGenerator):  84,9 /  76,1 / 201,4 msecs
Length    8,   end escape: (7-bit, 8-bit, JsonGenerator):  74,2 /  70,7 / 230,7 msecs
Length   16,  none escape: (7-bit, 8-bit, JsonGenerator):  65,2 /  56,1 / 121,2 msecs
Length   16, start escape: (7-bit, 8-bit, JsonGenerator):  74,0 /  62,3 / 133,6 msecs
Length   16,   end escape: (7-bit, 8-bit, JsonGenerator):  67,8 /  59,3 / 178,4 msecs
Length   32,  none escape: (7-bit, 8-bit, JsonGenerator):  62,9 /  52,7 /  87,1 msecs
Length   32, start escape: (7-bit, 8-bit, JsonGenerator):  67,0 /  56,4 / 106,5 msecs
Length   32,   end escape: (7-bit, 8-bit, JsonGenerator):  63,4 /  54,2 / 155,5 msecs
Length  256,  none escape: (7-bit, 8-bit, JsonGenerator):  60,2 /  52,2 /  61,7 msecs
Length  256, start escape: (7-bit, 8-bit, JsonGenerator):  63,2 /  54,1 /  86,0 msecs
Length  256,   end escape: (7-bit, 8-bit, JsonGenerator):  62,1 /  52,0 / 142,1 msecs
Length  512,  none escape: (7-bit, 8-bit, JsonGenerator):  59,1 /  51,1 /  56,9 msecs
Length  512, start escape: (7-bit, 8-bit, JsonGenerator):  62,2 /  53,1 /  82,6 msecs
Length  512,   end escape: (7-bit, 8-bit, JsonGenerator):  60,6 /  50,2 / 136,0 msecs
Length 1024,  none escape: (7-bit, 8-bit, JsonGenerator):  59,9 /  50,7 /  96,4 msecs
Length 1024, start escape: (7-bit, 8-bit, JsonGenerator):  60,6 /  52,8 /  81,3 msecs
Length 1024,   end escape: (7-bit, 8-bit, JsonGenerator):  60,5 /  49,7 /  97,0 msecs
Length 8192,  none escape: (7-bit, 8-bit, JsonGenerator):  59,1 /  48,9 /  45,1 msecs
Length 8192, start escape: (7-bit, 8-bit, JsonGenerator):  59,7 /  49,1 /  49,5 msecs
Length 8192,   end escape: (7-bit, 8-bit, JsonGenerator):  58,6 /  49,0 /  47,3 msecs


combined JsonGenerator results:

Length    8,  none escape: (before, after): 196,7  / 196,7 msecs
Length    8, start escape: (before, after): 201,4  / 198,5 msecs
Length    8,   end escape: (before, after): 230,7  / 231,6 msecs
Length   16,  none escape: (before, after): 121,2  / 120,0 msecs
Length   16, start escape: (before, after): 133,6  / 129,5 msecs
Length   16,   end escape: (before, after): 178,4  / 172,5 msecs
Length   32,  none escape: (before, after):  87,1  /  85,8 msecs
Length   32, start escape: (before, after): 106,5  / 103,6 msecs
Length   32,   end escape: (before, after): 155,5  / 146,5 msecs
Length  256,  none escape: (before, after):  61,7  /  60,7 msecs
Length  256, start escape: (before, after):  86,0  /  83,5 msecs
Length  256,   end escape: (before, after): 142,1  / 138,7 msecs
Length  512,  none escape: (before, after):  56,9  /  56,5 msecs
Length  512, start escape: (before, after):  82,6  /  79,3 msecs
Length  512,   end escape: (before, after): 136,0  / 132,9 msecs
Length 1024,  none escape: (before, after):  96,4  /  95,5 msecs
Length 1024, start escape: (before, after):  81,3  /  78,7 msecs
Length 1024,   end escape: (before, after):  97,0  /  97,0 msecs
Length 8192,  none escape: (before, after):  45,1  /  32,4 msecs
Length 8192, start escape: (before, after):  49,5  /  38,1 msecs
Length 8192,   end escape: (before, after):  47,3  /  34,6 msecs
```

It's not as big of a difference I was seeing in the original benchmarks I had, although it's noticeable how larger strings appear to benefit quite a bit.

---

There is potentially another question of whether `UTF8JsonGenerator#_extendOutputEscapesTo8Bits` should overwrite `JsonGeneratorImpl#sOutputEscapes`, as to avoid having to recreate the 8-bit wide LUT for each individual `JsonGenerator` instance. I opted not to change this since that crosses `UTF8JsonGenerator`'s boundary into the parent class, as well as demoting `JsonGeneratorImpl#sOutputEscapes` to a non-final field, which feels iffy.