@cowtowncoder 2024-10-27
@JoostK First of all: thank you for contributing this! At high level this makes sense, but I do need to dig bit deeper into this when I have time -- and right now I am bit overloaded/overspread so apologies for delay there may be.

Having said that: one thing we will eventually need (if not done; apologies if it has) is to get a CLA, from here:

https://github.com/FasterXML/jackson/blob/master/contributor-agreement.pdf

(needs to be sent just once for all Jackson contributions)

the usual way is to print, fill & sign, scan/photo, email to "cla" at fasterxml dot com.
Once I receive it we are good to wrt merging (obv pending code review).

Thank you again; looking forward to getting this merged!